Locations clickable for availability and appointment
I think I covered all the bases. Online campus locations now show up as a link. If they aren't online, it will just show up as plain text. Changes affect the popups on index for both clients and provider when clicking on appointments (also affects popups for availability for provider). Also added campus to the popups that didn't have it. Links are not clickable on registering for an appointment, Links are only clickable for clients after they've registered for the appointment.